+++ Cinema/MovieTicket.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Text.Json.Serialization;
 
-namespace cinema_sofa
+namespace Cinema
 {
     public class MovieTicket
     {
